test(client-transcribe-streaming): add concurrency to startStreamTranscription test
