+++ clients/client-transcribe-streaming/package.json
@@ -13,14 +13,14 @@
     "extract:docs": "api-extractor run --local",
     "generate:client": "node ../../scripts/generate-clients/single-service --solo transcribe-streaming",
     "test": "yarn g:vitest run --passWithNoTests",
+    "test:watch": "yarn g:vitest watch --passWithNoTests",
     "test:browser": "yarn g:vitest run -c vitest.config.browser.e2e.mts",
     "test:browser:watch": "yarn g:vitest watch -c vitest.config.browser.e2e.mts",
     "test:e2e": "yarn g:vitest run -c vitest.config.e2e.mts --mode development",
     "test:e2e:watch": "yarn g:vitest watch -c vitest.config.e2e.mts",
     "test:index": "tsc --noEmit ./test/index-types.ts && node ./test/index-objects.spec.mjs",
     "test:integration": "yarn g:vitest run --passWithNoTests -c vitest.config.integ.mts",
-    "test:integration:watch": "yarn g:vitest run --passWithNoTests -c vitest.config.integ.mts",
-    "test:watch": "yarn g:vitest watch --passWithNoTests"
+    "test:integration:watch": "yarn g:vitest run --passWithNoTests -c vitest.config.integ.mts"
   },
   "main": "./dist-cjs/index.js",
   "types": "./dist-types/index.d.ts",
